+++ index.js
@@ -36,3 +36,4 @@
 
 export {default} from './lib/router.js';
 export {default as graceful} from './lib/lifecycle.js';
+export {presets} from './lib/presets.js';